refactor: extract getNodesByGitignorePath
refs: #2318
