+++ visualization/app/codeCharta/state/selectors/searchedNodes/getNodesByGitignorePath.ts
@@ -1,0 +1,19 @@
+import { hierarchy } from "d3-hierarchy"
+
+import { CodeMapNode } from "../../../codeCharta.model"
+import { returnIgnore, transformPath } from "../../../util/codeMapHelper"
+
+export function getNodesByGitignorePath(root: CodeMapNode, gitignorePath: string) {
+	gitignorePath = gitignorePath.trimStart()
+	if (gitignorePath.length === 0) {
+		return []
+	}
+	const ignoreResults = returnIgnore(gitignorePath)
+	const filtered = []
+	for (const { data } of hierarchy(root)) {
+		if (ignoreResults.ignoredNodePaths.ignores(transformPath(data.path)) === ignoreResults.condition) {
+			filtered.push(data)
+		}
+	}
+	return filtered
+}
